Atualizacao de limite do cartao

+++ creditcard/src/main/java/com/github/rafaelfernandes/creditcard/adapter/in/web/response/CreditCardDataResponse.java
@@ -1,5 +1,6 @@
 package com.github.rafaelfernandes.creditcard.adapter.in.web.response;
 
+import java.math.BigDecimal;
 import java.math.BigInteger;
 
 public record CreditCardDataResponse(
@@ -7,6 +8,6 @@ public record CreditCardDataResponse(
         String numero,
         String data_validade,
         String cvv,
-        BigInteger limite
+        BigDecimal limite
 ) {
 }
